move demo form to subfolder
