+++ redux-form-tutorial/src/main/typescript/lib/form.tsx
@@ -2,17 +2,19 @@ import * as React from 'react'
 import {WrappedFieldInputProps} from "redux-form"
 import {BaseFieldProps, WrappedFieldProps} from "redux-form/lib/Field"
 
-type FieldProps = WrappedFieldProps & BaseFieldProps & {submittingFailed?: boolean}
+type FieldProps = WrappedFieldProps & BaseFieldProps
 type renderer = (input: WrappedFieldInputProps, label?: string, rest?: any) => JSX.Element
 
-const createRenderer = (renderer: renderer) => ({input, meta, label, submittingFailed, ...rest}: FieldProps) => {
+const createRenderer = (renderer: renderer) => ({input, meta, label, ...rest}: FieldProps) => {
+    const errorCondition = meta.error && meta.touched
+
     return <div className={[
-        meta.error ? 'error' : '',
+        errorCondition ? 'error' : '',
         meta.active ? 'active' : '',
     ].join(' ')}>
         <label>{label}</label>
         {renderer(input, label, rest)}
-        {meta.error && <span>{meta.error}</span>}
+        {errorCondition && <span>{meta.error}</span>}
     </div>
 }
